Fix ModuleManager configs for better compatibility.
ModuleManager documentation, what exists of it, recommends that mod authors do
not use FINAL: this is reserved for the end user to have control over the
final state of the configs without deleting or altering supplied files.

Instead, it is recommended to mark patches with FOR[], and maintain patch
ordering with AFTER[], which is what I've done here.

Also, CargoShroud is a part (nosecone in Tundra kit) and CargoShorud is not a
typo, it's another part (decoupling solar panel cover for Cargo Dragon) --
which also has no business getting a decoupler shroud.